feat: implement multi-pass wave-style path tracing
- Replace single CSMain kernel with 4 kernels: generate, trace_shade, finalize, transfer
- Add indirect dispatch for active ray compression per bounce
- Implement ping-pong double buffering (GlobalRaysA/B) to avoid read-write races
- Add TransferKernel to compute indirect dispatch args from BufferSizes counter
- Add active ray compression via InterlockedAdd on next bounce counter
- Fix debug mode output (write to GlobalColors instead of _Result)
- Fix RNG shadow bug (use global rng in kernel_generate)
- Fix IndirectArgs buffer stride (uint3->uint, ComputeBuffer(3,4))
- Fix PathContribution stride alignment (pad to 32 bytes)
- Fix NeedUpdate flag (initial true, clear after SetShaderParameters)
- Fix LightCullingManager hardcoded kernel 0 binding
- Fix TransferKernel Type default value (int Type = 0)
- Add light buffer binding overloads for multi-kernel support
- Add debug mode skip optimization (skip bounce loop)
- Optimize GC: reuse CommandBuffer, cache kernel index arrays, pre-compute bounce names
- Optimize GC: use stackalloc in OnDrawGizmos, reuse sizesData array
- Fix RNG_Next return type half->float, SampleHemisphere half->float

+++ Assets/ComputeShader/main/function.hlsl
@@ -63,10 +63,10 @@ void RNG_SeedPixel(inout RNG rng, uint2 pixelXY, uint frameId /*帧序号*/)
     rng.state = WangHash(s);
 }
 
-half RNG_Next(inout RNG rng)  // float -> half
+float RNG_Next(inout RNG rng)
 {
     rng.state = WangHash(rng.state);
-    return half(HashToFloat(rng.state));
+    return HashToFloat(rng.state);
 }
 
 float SmoothnessToPhongAlpha(float s)
@@ -99,16 +99,16 @@ float3 SampleReflectionDirectionSphere(float3 normal, float alpha)
     return mul(tangentSpaceDir, GetTangentSpace(normal));
 }
 
-half3 SampleHemisphere(float3 norm)
+float3 SampleHemisphere(float3 norm)
 {
-    half u1 = RNG_Next(rng);
-    half u2 = RNG_Next(rng);
+    float u1 = RNG_Next(rng);
+    float u2 = RNG_Next(rng);
 
-    half r     = sqrt(u1);
-    half theta = 2 * PI * u2;
-    half x     = r * cos(theta);
-    half y     = r * sin(theta);
-    half z     = sqrt(max(0, 1 - u1));
+    float r     = sqrt(u1);
+    float theta = 2 * PI * u2;
+    float x     = r * cos(theta);
+    float y     = r * sin(theta);
+    float z     = sqrt(max(0, 1 - u1));
 
     float3 tangent   = normalize(cross(abs(norm.z) < 0.999 ? float3(0,0,1) : float3(1,0,0), norm));
     float3 bitangent = cross(norm, tangent);
